refactor: move header to separate component
